refactor to new single page
- remove redundant files
- clean up file structure
- clean up grid spacing to remove horizontal scrolling and padding
- move config over to apiv2
- some linting
