+++ components/About.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import Layout from '~/components/Layout';
 import { Grid, Typography } from '@material-ui/core';
 
 const About = () => (
